@@ -1,10 +1,8 @@
-const sharp = require('sharp');
-const fs = require('fs').promises;
-const path = require('path');
+import sharp from 'sharp';
+import fs from 'fs';
+import path from 'path';
 
-const IMAGES_DIR = path.join(process.cwd(), 'public', 'images');
-
-async function convertToAvif(inputPath, outputPath) {
+const convertToAvif = async (inputPath, outputPath) => {
   try {
     await sharp(inputPath)
       .avif({ quality: 80, effort: 6 })
@@ -13,34 +11,27 @@ async function convertToAvif(inputPath, outputPath) {
   } catch (error) {
     console.error(`Error converting ${inputPath}:`, error);
   }
-}
+};
 
-async function processDirectory(dir) {
-  const entries = await fs.readdir(dir, { withFileTypes: true });
-
-  for (const entry of entries) {
-    const fullPath = path.join(dir, entry.name);
+const processDirectory = async (dir) => {
+  const files = fs.readdirSync(dir);
+  
+  for (const file of files) {
+    const fullPath = path.join(dir, file);
+    const stat = fs.statSync(fullPath);
     
-    if (entry.isDirectory()) {
+    if (stat.isDirectory()) {
       await processDirectory(fullPath);
-      continue;
-    }
-
-    if (entry.isFile() && /\.(jpg|jpeg|png)$/i.test(entry.name)) {
-      const avifPath = fullPath.replace(/\.(jpg|jpeg|png)$/i, '.avif');
-      await convertToAvif(fullPath, avifPath);
+    } else if (file.match(/\.(jpg|jpeg|png)$/i)) {
+      const outputPath = fullPath.replace(/\.(jpg|jpeg|png)$/i, '.avif');
+      await convertToAvif(fullPath, outputPath);
     }
   }
-}
+};
 
-async function main() {
-  try {
-    await processDirectory(IMAGES_DIR);
-    console.log('Image conversion complete!');
-  } catch (error) {
-    console.error('Error processing images:', error);
-    process.exit(1);
-  }
-}
+const start = async () => {
+  const publicDir = path.join(process.cwd(), 'public', 'images');
+  await processDirectory(publicDir);
+};
 
-main(); 
+start().catch(console.error); 
